aider: Add "openshift-cherrypick-robot" to "org-1" and "org-3" in fakeAutomationClient.
# Aider chat conversation:

USER: Add openshift-cherrypick-robot to some of the organizations in your fakeAutomationClient:

membersByOrg: map[string][]string{
    "org-1": {"a-user", "d-bot", "e-bot", "openshift-cherrypick-robot"},
    "org-2": {"some-user", "z-bot"},
    "org-3": {"a-user", "openshift-cherrypick-robot"},
},

ASSISTANT: To add "openshift-cherrypick-robot" to the organizations "org-1" and "org-3" in the fakeAutomationClient, we need to modify the "main_test.go" file.

Here is the *edit block* for the change:

```go
main_test.go
<<<<<<< HEAD
	membersByOrg: map[string][]string{
		"org-1": {"a-user", "d-bot", "e-bot"},
		"org-2": {"some-user", "z-bot"},
		"org-3": {"a-user"},
	},
=======
	membersByOrg: map[string][]string{
		"org-1": {"a-user", "d-bot", "e-bot", "openshift-cherrypick-robot"},
		"org-2": {"some-user", "z-bot"},
		"org-3": {"a-user", "openshift-cherrypick-robot"},
	},
>>>>>>> updated
```
This change adds "openshift-cherrypick-robot" to the list of members for "org-1" and "org-3".