aider: Extend automationClient interface to check if a repo has the cherrypick external prow plugin configured, and then check if openshift-cherrypick-robot is an org member.
# Aider chat conversation:

USER: Conditional Check for Cherrypick Plugin: Add logic to determine if a repo has the cherrypick external prow plugin configured. Only then proceed to check if openshift-cherrypick-robot is an org member.
You may need to extend the automationClient interface or utilize another data source to check for this plugin configuration.

ASSISTANT: To implement this feature, we need to extend the `automationClient` interface to include a method that checks if a repo has the cherrypick external prow plugin configured. Let's call this method `IsCherrypickPluginConfigured(org, repo string) (bool, error)`.

We will then use this method in the `checkRepos` function to determine if we should check if `openshift-cherrypick-robot` is an org member.

Here are the changes we need to make:

1. Extend the `automationClient` interface in `main.go` to include the `IsCherrypickPluginConfigured` method.
2. Update the `checkRepos` function in `main.go` to use the `IsCherrypickPluginConfigured` method.

Here are the edit blocks for these changes:

```go
main.go
<<<<<<< HEAD
type automationClient interface {
	IsMember(org, user string) (bool, error)
	IsCollaborator(org, repo, user string) (bool, error)
	IsAppInstalled(org, repo string) (bool, error)
}
=======
type automationClient interface {
	IsMember(org, user string) (bool, error)
	IsCollaborator(org, repo, user string) (bool, error)
	IsAppInstalled(org, repo string) (bool, error)
	IsCherrypickPluginConfigured(org, repo string) (bool, error)
}
>>>>>>> updated
```

```go
main.go
<<<<<<< HEAD
		// Add a check for openshift-cherrypick-robot
		isCherryPickRobotMember, err := client.IsMember(org, "openshift-cherrypick-robot")
		if err != nil {
			return nil, fmt.Errorf("unable to determine if openshift-cherrypick-robot is a member of %s: %w", org, err)
		}
		if !isCherryPickRobotMember {
			failing.Insert(orgRepo)
			repoLogger.Error("openshift-cherrypick-robot is not an org member")
		} else {
			repoLogger.Info("openshift-cherrypick-robot is an org member")
		}
=======
		// Check if the cherrypick plugin is configured for the repo
		isCherrypickPluginConfigured, err := client.IsCherrypickPluginConfigured(org, repo)
		if err != nil {
			return nil, fmt.Errorf("unable to determine if cherrypick plugin is configured for %s/%s: %w", org, repo, err)
		}
		if isCherrypickPluginConfigured {
			// Add a check for openshift-cherrypick-robot
			isCherryPickRobotMember, err := client.IsMember(org, "openshift-cherrypick-robot")
			if err != nil {
				return nil, fmt.Errorf("unable to determine if openshift-cherrypick-robot is a member of %s: %w", org, err)
			}
			if !isCherryPickRobotMember {
				failing.Insert(orgRepo)
				repoLogger.Error("openshift-cherrypick-robot is not an org member")
			} else {
				repoLogger.Info("openshift-cherrypick-robot is an org member")
			}
		}
>>>>>>> updated
```

Please note that the implementation of the `IsCherrypickPluginConfigured` method depends on the specific way your system is set up to configure plugins. You will need to implement this method in the concrete type that implements the `automationClient` interface.